fix: Pattern Strategy para Scraper e Templethod Method para Verificador
