+++ python/scraper/settings.py
@@ -1,0 +1,52 @@
+DB_NAME = '../../database/database.sqlite'
+
+urls_para_processar = [
+    {"url": "https://patoloco.com.br/produtos/placa-de-video", "filters": None},
+    {"url": "https://patoloco.com.br/produtos/processadores", "filters": None},
+    {"url": "https://patoloco.com.br/produtos/placas-mae", "filters": None},
+    {"url": "https://patoloco.com.br/produtos/memorias", "filters": None},
+    {"url": "https://patoloco.com.br/produtos/ssd", "filters": None},
+    {"url": "https://patoloco.com.br/produtos/ssd-m2", "filters": None},
+    {"url": "https://patoloco.com.br/produtos/coolers-e-watercoolers", "filters": None},
+    {"url": "https://patoloco.com.br/produtos/fontes", "filters": None},
+    {"url": "https://www.kabum.com.br/hardware/placa-de-video-vga",
+     "filters": {
+         "facet_filters": "eyJrYWJ1bV9wcm9kdWN0IjpbInRydWUiXSwiY2F0ZWdvcnkiOlsiSGFyZHdhcmUiXX0=",
+         "sort": "most_searched"
+     }},
+    {"url": "https://www.kabum.com.br/hardware/processadores",
+         "filters": {
+             "facet_filters": "eyJrYWJ1bV9wcm9kdWN0IjpbInRydWUiXSwiY2F0ZWdvcnkiOlsiSGFyZHdhcmUiXSwiVsOtZGVvIEludGVncmFkbyI6WyJOw6NvIl19",
+             "sort": "most_searched"
+         }},
+        {"url": "https://www.kabum.com.br/hardware/ssd-2-5/ssd-sata",
+             "filters": {
+                 "facet_filters": "eyJrYWJ1bV9wcm9kdWN0IjpbInRydWUiXSwiY2F0ZWdvcnkiOlsiSGFyZHdhcmUiXSwiaGFzX29mZmVyIjpbInRydWUiXX0=",
+                 "sort": "most_searched"
+             }},
+        {"url": "https://www.kabum.com.br/hardware/coolers/water-cooler",
+             "filters": {
+                 "facet_filters": "eyJrYWJ1bV9wcm9kdWN0IjpbInRydWUiXSwiY2F0ZWdvcnkiOlsiSGFyZHdhcmUiXX0=",
+                 "sort": "most_searched"
+             }},
+        {"url": "https://www.kabum.com.br/hardware/coolers/air-cooler",
+             "filters": {
+                 "facet_filters": "eyJrYWJ1bV9wcm9kdWN0IjpbInRydWUiXSwiY2F0ZWdvcnkiOlsiSGFyZHdhcmUiXX0=",
+                 "sort": "most_searched"
+             }},
+        {"url": "https://www.kabum.com.br/hardware/fontes",
+             "filters": {
+                 "facet_filters": "eyJrYWJ1bV9wcm9kdWN0IjpbInRydWUiXSwiY2F0ZWdvcnkiOlsiSGFyZHdhcmUiXX0=",
+                 "sort": "most_searched"
+             }},
+        {"url": "https://www.kabum.com.br/hardware/memoria-ram",
+             "filters": {
+                 "facet_filters": "eyJrYWJ1bV9wcm9kdWN0IjpbInRydWUiXSwiY2F0ZWdvcnkiOlsiSGFyZHdhcmUiXSwiQ29tcGF0aWJpbGlkYWRlIjpbIkRlc2t0b3AiXSwiaGFzX29mZmVyIjpbInRydWUiXX0=",
+                 "sort": "most_searched"
+             }},
+        {"url": "https://www.kabum.com.br/hardware/placas-mae",
+             "filters": {
+                 "facet_filters": "eyJrYWJ1bV9wcm9kdWN0IjpbInRydWUiXSwiaGFzX29mZmVyIjpbInRydWUiXX0=",
+                 "sort": "most_searched"
+             }},
+]
